# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Transaction Microservice API
  version: 1.0.0
  description: API para gestionar transacciones bancarias como depósitos, retiros, pagos, transferencias y cargos por crédito.

servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo

paths:
  /api/transactions:
    get:
      summary: Listar todas las transacciones
      operationId: getAllTransactions
      responses:
        '200':
          description: Lista de transacciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'

    post:
      summary: Crear una nueva transacción
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '200':
          description: Transacción registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/transactions/{id}:
    get:
      summary: Obtener transacción por ID
      operationId: getTransactionById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transacción encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Transacción no encontrada

    put:
      summary: Actualizar transacción por ID
      operationId: updateTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '200':
          description: Transacción actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Transacción no encontrada

    delete:
      summary: Eliminar transacción por ID
      operationId: deleteTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transacción eliminada
        '404':
          description: Transacción no encontrada

  /api/transactions/transfer:
    post:
      summary: Realizar transferencia entre cuentas
      description: >
        Este endpoint está protegido con Circuit Breaker y TimeLimiter usando Resilience4j.
        Si el servicio externo tarda más de 2 segundos en responder, se activa el timeout y se retorna una respuesta fallback.
        El circuito se abre tras múltiples fallos consecutivos para evitar sobrecarga.

      operationId: transferBetweenAccounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transferencia registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'

  /api/transactions/commissions:
    get:
      summary: Obtener comisiones cobradas por producto en un periodo
      operationId: getCommissionsByProduct
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista de comisiones por producto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommissionReport'

  /api/transactions/daily-balances:
    get:
      summary: Obtener saldos diarios por cliente en un mes
      operationId: getDailyBalances
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: string
        - name: month
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        '200':
          description: Lista de saldos diarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyBalance'

components:
  schemas:
    TransactionRequest:
      type: object
      required:
        - productId
        - type
        - amount
        - date
        - accountId
        - operationTypeId
        - status
        - origin
        - eventDate
      properties:
        productId:
          type: string
        customerId:
          type: string
        accountId:
          type: string
        operationTypeId:
          type: string
        type:
          type: string
          enum: [ DEPOSIT, WITHDRAWAL, PAYMENT, CREDIT_CHARGE, TRANSFER_INTERNAL, TRANSFER_EXTERNAL ]
        origin:
          type: string
          enum: [ INTERNAL, EXTERNAL ]
        amount:
          type: number
          format: decimal
          minimum: 0.01
        date:
          type: string
          format: date-time
        eventDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [ APPROVED, REJECTED, PENDING ]
        description:
          type: string
          nullable: true
        destinationAccountId:
          type: string
          nullable: true
        commissionApplied:
          type: number
          format: decimal
          nullable: true

    TransactionResponse:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        customerId:
          type: string
        accountId:
          type: string
        operationTypeId:
          type: string
        type:
          type: string
          enum: [ DEPOSIT, WITHDRAWAL, PAYMENT, CREDIT_CHARGE, TRANSFER_INTERNAL, TRANSFER_EXTERNAL ]
        origin:
          type: string
          enum: [ INTERNAL, EXTERNAL ]
        amount:
          type: number
          format: decimal
        date:
          type: string
          format: date-time
        eventDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [ APPROVED, REJECTED, PENDING ]
        description:
          type: string
          nullable: true
        destinationAccountId:
          type: string
          nullable: true
        commissionApplied:
          type: number
          format: decimal
          nullable: true

    TransferRequest:
      type: object
      required:
        - sourceAccountId
        - destinationAccountId
        - amount
        - customerId
      properties:
        sourceAccountId:
          type: string
        destinationAccountId:
          type: string
        amount:
          type: number
          format: decimal
        customerId:
          type: string

    CommissionReport:
      type: object
      properties:
        productId:
          type: string
        productType:
          type: string
        totalCommission:
          type: number
          format: decimal

    DailyBalance:
      type: object
      properties:
        date:
          type: string
          format: date
        balance:
          type: number
          format: decimal